+++ Shikkhanobish/Shikkhanobish/Model/Teacher.cs
@@ -25,7 +25,7 @@
         public int Age { get; set; }
         public string Class { get; set; }
         public string InstitutionName { get; set; }
-        public int RechargedAmount { get; set; }
+        public float RechargedAmount { get; set; }
         public int isFounder { get; set; }
         public string Response { get; set; }
         public double Amount { get; set; }
